+++ src/framework/piral-core/src/types/extension.ts
@@ -1,5 +1,6 @@
 import type { ReactNode, ReactElement } from 'react';
 import type { PiralCustomExtensionSlotMap } from './custom';
+import type { ExtensionRegistration } from './state';
 
 /**
  * The mapping of the existing (known) extension slots.
@@ -20,8 +21,26 @@ export interface BaseExtensionSlotProps<TName, TParams> {
    */
   empty?(): ReactNode;
   /**
+   * Determines if the `render` function should be called in case no
+   * components are available for the specified extension.
+   *
+   * If true, `empty` will be called and returned from the slot.
+   * If false, `render` will be called with the result of calling `empty`.
+   * The result of calling `render` will then be returned from the slot.
+   */
+  emptySkipsRender?: boolean;
+  /**
+   * Defines the order of the components to render.
+   * May be more convient than using `render` w.r.t. ordering extensions
+   * by their supplied metadata.
+   * @param extensions The registered extensions.
+   * @returns The ordered extensions.
+   */
+  order?(extensions: Array<ExtensionRegistration>): Array<ExtensionRegistration>;
+  /**
    * Defines how the provided nodes should be rendered.
    * @param nodes The rendered extension nodes.
+   * @returns The rendered nodes, i.e., an ReactElement.
    */
   render?(nodes: Array<ReactNode>): ReactElement<any, any> | null;
   /**