+++ src/tooling/piral-cli/src/common/clients/npm.ts
@@ -1,13 +1,18 @@
 import { resolve } from 'path';
 import { log } from '../log';
+import { findFile } from '../io';
 import { runCommand } from '../scripts';
 import { MemoryStream } from '../MemoryStream';
+
+// Helpers:
 
 function runNpmProcess(args: Array<string>, target: string, output?: NodeJS.WritableStream) {
   log('generalDebug_0003', 'Starting the npm process ...');
   const cwd = resolve(process.cwd(), target);
   return runCommand('npm', args, cwd, output);
 }
+
+// Client interface functions:
 
 export async function installDependencies(target = '.', ...flags: Array<string>) {
   const ms = new MemoryStream();
@@ -16,26 +21,37 @@ export async function installDependencies(target = '.', ...flags: Array<string>)
   return ms.value;
 }
 
+export async function installPackage(packageRef: string, target = '.', ...flags: Array<string>) {
+  const ms = new MemoryStream();
+  await runNpmProcess(['install', packageRef, '--legacy-peer-deps', ...flags], target, ms);
+  log('generalDebug_0003', `npm install package result: ${ms.value}`);
+  return ms.value;
+}
+
+export async function detectClient(root: string) {
+  return !!(await findFile(root, 'package-lock.json'));
+}
+
+export async function isProject(root: string, packageRef: string) {
+  const details = await listPackage(packageRef, root);
+  const packageDetails =  details?.dependencies?.[packageRef];
+
+  if (packageDetails && typeof packageDetails.resolved === 'string') {
+    return packageDetails.resolved.startsWith('file:');
+  }
+
+  return false;
+}
+
+export async function initProject(projectName: string, target: string) {}
+
+// Functions to exclusively use from npm client:
+
 export async function unpackPackage(packageRef: string, target = '.', ...flags: Array<string>) {
   const ms = new MemoryStream();
   await runNpmProcess(['pack', packageRef, ...flags], target, ms);
   log('generalDebug_0003', `npm (un)pack result: ${ms.value}`);
   return ms.value;
-}
-
-export async function installPackage(packageRef: string, target = '.', ...flags: Array<string>) {
-  try {
-    const ms = new MemoryStream();
-    await runNpmProcess(['install', packageRef, '--legacy-peer-deps', ...flags], target, ms);
-    log('generalDebug_0003', `npm install package result: ${ms.value}`);
-    return ms.value;
-  } catch (ex) {
-    log(
-      'generalError_0002',
-      `Could not install the package "${packageRef}" using npm. Make sure npm is correctly installed and accessible: ${ex}`,
-    );
-    throw ex;
-  }
 }
 
 export async function createPackage(target = '.', ...flags: Array<string>) {
@@ -73,6 +89,7 @@ export async function listPackage(packageRef: string, target = '.', ...flags: Ar
     await runNpmProcess(['ls', packageRef, '--json', '--depth', '0', ...flags], target, ms);
   } catch (e) {
     log('generalDebug_0003', `npm ls packageRef error: ${e}`);
+    return {};
   }
 
   log('generalDebug_0003', `npm ls packageRef result: ${ms.value}`);