+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed issue with retrieving type root
 - Fixed `exports` field to use `*` notation
+- Added `noEmptyRender` prop to `ExtensionSlot`
 
 ## 0.14.23 (May 1, 2022)