+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Piral Changelog
 
+## 0.14.24 (tbd)
+
+- Fixed issue with retrieving type root
+- Fixed `exports` field to use `*` notation
+- Fixed resolution of `piral-cli` plugins with pnpm (#514)
+- Fixed missing `PIRAL_PUBLIC_PATH` when building a Piral instance with `piral-cli-esbuild`
+- Added `emptySkipsRender` prop to `ExtensionSlot`
+- Added `order` prop to `ExtensionSlot`
+
 ## 0.14.23 (May 1, 2022)
 
 - Improved typings for registering extension components